add more code in calc_delta_dist_filter_boundary

+++ src/enc/pinter.rs
@@ -643,8 +643,8 @@ impl EvceCtx {
                 }
 
                 self.calc_delta_dist_filter_boundary(
-                    x,
-                    y,
+                    x as i16,
+                    y as i16,
                     log2_cuw,
                     log2_cuh,
                     self.core.avail_lr,
@@ -1094,8 +1094,8 @@ impl EvceCtx {
                 let refi = self.pinter.refi[pidx];
                 let mv = self.pinter.mv[pidx];
                 self.calc_delta_dist_filter_boundary(
-                    x,
-                    y,
+                    x as i16,
+                    y as i16,
                     log2_cuw,
                     log2_cuh,
                     self.core.avail_lr,
@@ -1160,8 +1160,8 @@ impl EvceCtx {
                 let refi = self.pinter.refi[pidx];
                 let mv = self.pinter.mv[pidx];
                 self.calc_delta_dist_filter_boundary(
-                    x,
-                    y,
+                    x as i16,
+                    y as i16,
                     log2_cuw,
                     log2_cuh,
                     self.core.avail_lr,
@@ -1380,8 +1380,8 @@ impl EvceCtx {
                 let refi = self.pinter.refi[pidx];
                 let mv = self.pinter.mv[pidx];
                 self.calc_delta_dist_filter_boundary(
-                    x,
-                    y,
+                    x as i16,
+                    y as i16,
                     log2_cuw,
                     log2_cuh,
                     self.core.avail_lr,
